+++ again/app/src/main/java/org/sopt/study/againassignment/etc/ViewExt.kt
@@ -1,0 +1,8 @@
+package org.sopt.study.againassignment
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.shortToast(message:String){
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
