+++ again/app/src/main/java/org/sopt/study/againassignment/server/SignUp.kt
@@ -1,0 +1,16 @@
+package org.sopt.study.againassignment.server
+
+import org.sopt.study.againassignment.data.RequestSignupData
+import org.sopt.study.againassignment.data.ResponseSignupData
+import retrofit2.Call
+import retrofit2.http.Body
+import retrofit2.http.Headers
+import retrofit2.http.POST
+
+interface SignUp {
+    @Headers("Content-Type:application/json")
+    @POST("user/signup")
+    fun postSignup(
+        @Body body : RequestSignupData
+    ) : Call<ResponseSignupData>
+}
